Move prime metrics to inference
